transferred sample frame image generation into the body of transform_and_save_dataset and transform_two_chunks
Sample frame image generation used to be done within generate_report_from_xls_registration_summary.
That had problems - especially when the summary stack was not saved.
A lot of data need to be carefully kept in order for correct evaluation box to be displayed.
Now the sample frame images with evaluation box superimposed are created by the same functions that perform transformation and evaluation - easier to keep track and debug.